checkpoint, docs fixed
