+++ crates/steel-core/src/compiler/compiler.rs
@@ -17,7 +17,11 @@ use crate::{
         kernel::Kernel,
         parser::{lower_entire_ast, lower_macro_and_require_definitions},
     },
-    steel_vm::{builtin::BuiltInModule, cache::MemoizationTable, engine::ModuleContainer},
+    steel_vm::{
+        builtin::BuiltInModule,
+        cache::MemoizationTable,
+        engine::{ModuleContainer, DEFAULT_DOC_MACROS},
+    },
 };
 use crate::{
     core::{instructions::Instruction, opcode::OpCode},
@@ -469,7 +473,10 @@ impl Compiler {
 
         // Could fail here
         let parsed: std::result::Result<Vec<ExprKind>, ParseError> =
-            Parser::new(expr_str, Some(id)).collect();
+            Parser::new(expr_str, Some(id))
+                .without_lowering()
+                .map(|x| x.and_then(lower_macro_and_require_definitions))
+                .collect();
 
         let parsed = parsed?;
 
@@ -649,7 +656,6 @@ impl Compiler {
             // are safe to eliminate. In interactive mode, we'll
             // be unable to optimize those away
             .remove_unused_globals_with_prefix("mangler", &self.macro_env, &self.module_manager);
-
         // Don't do lambda lifting here
         // .lift_pure_local_functions()
         // .lift_all_local_functions();
@@ -713,10 +719,10 @@ impl Compiler {
         let mut expanded_statements =
             self.expand_expressions(exprs, path, sources, builtin_modules.clone())?;
 
-        expanded_statements = expanded_statements
-            .into_iter()
-            .map(lower_entire_ast)
-            .collect::<std::result::Result<Vec<_>, ParseError>>()?;
+        // expanded_statements = expanded_statements
+        //     .into_iter()
+        //     .map(lower_entire_ast)
+        //     .collect::<std::result::Result<Vec<_>, ParseError>>()?;
 
         // if log_enabled!(log::Level::Debug) {
         //     debug!(
@@ -753,7 +759,21 @@ impl Compiler {
             .map(lower_entire_ast)
             .collect::<std::result::Result<Vec<_>, ParseError>>()?;
 
-        // expanded_statements.pretty_print();
+        expanded_statements = expanded_statements
+            .into_iter()
+            .map(|x| {
+                expand_kernel_in_env(
+                    x,
+                    self.kernel.as_mut(),
+                    builtin_modules.clone(),
+                    "top-level".to_string(),
+                )
+            })
+            .collect::<Result<Vec<_>>>()?;
+
+        // println!("{:#?}", expanded_statements);
+
+        // Let's do the @doc macro here?
 
         log::debug!(target: "expansion-phase", "Beginning constant folding");
 
@@ -796,7 +816,13 @@ impl Compiler {
         analysis.fresh_from_exprs(&expanded_statements);
         analysis.populate_captures(&expanded_statements);
 
+        // expanded_statements.pretty_print();
+
         let mut semantic = SemanticAnalysis::from_analysis(&mut expanded_statements, analysis);
+
+        // Check if the values are redefined!
+        semantic.check_if_values_are_redefined()?;
+
         semantic.refresh_variables();
         semantic.flatten_anonymous_functions();
         semantic.refresh_variables();
