+++ crates/steel-core/src/scheme/modules/result.scm
@@ -15,9 +15,9 @@
          Result/c
          ; (contract/out unwrap-ok (->/c Ok? any/c))
          unwrap-ok
-         (contract/out unwrap-err (->/c Err? any/c))
-         (contract/out map-ok (->/c Result? (->/c any/c any/c) Result?))
-         (contract/out map-err (->/c Result? (->/c any/c any/c) Result?))
+         (contract/out/test unwrap-err (->/c Err? any/c))
+         (contract/out/test map-ok (->/c Result? (->/c any/c any/c) Result?))
+         (contract/out/test map-err (->/c Result? (->/c any/c any/c) Result?))
          unwrap-or
          ok-and-then)
