+++ crates/steel-core/src/steel_vm/engine.rs
@@ -278,6 +278,13 @@ macro_rules! time {
     }};
 }
 
+thread_local! {
+    pub(crate) static DEFAULT_PRELUDE_MACROS: RefCell<HashMap<InternedString, SteelMacro>> = RefCell::new(HashMap::new());
+
+    pub(crate) static DEFAULT_DOC_MACROS: RefCell<HashMap<InternedString, SteelMacro>> = RefCell::new(
+        HashMap::new());
+}
+
 impl Engine {
     /// Function to access a kernel level execution environment
     /// Has access to primitives and syntax rules, but will not defer to a child
@@ -318,6 +325,33 @@ impl Engine {
         for core in core_libraries.into_iter() {
             vm.compile_and_run_raw_program(core).unwrap();
         }
+
+        // vm.compile_and_run_raw_program(
+        //     r#"
+
+        // (define-syntax @doc
+        //     (syntax-rules ()
+        //         [(@doc comment expression)
+        //         expression]))
+        //     "#,
+        // )
+        // .unwrap();
+
+        // let at_doc = "@doc".into();
+
+        // DEFAULT_DOC_MACROS.with(|x| {
+        //     x.borrow_mut()
+        //         .insert(at_doc, vm.in_scope_macros_mut().remove(&at_doc).unwrap())
+        // });
+
+        // Initialize the global macro environment with the default one. This way
+        // values won't leak when top level macros are defined - and modules can clone from
+        // this to begin seeding their environment.
+        DEFAULT_PRELUDE_MACROS.with(|x| {
+            let mut guard = x.borrow_mut();
+
+            *guard = vm.in_scope_macros().clone();
+        });
 
         log::debug!(target: "kernel", "Loaded prelude in the kernel!: {:?}", now.elapsed());
