+++ crates/steel-language-server/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod backend;
 mod client;
 mod configuration;
 mod contract;
