+++ r7rs-benchmarks/simplex.scm
@@ -14,7 +14,7 @@
 
 (define vector-ref mut-vector-ref)
 ; (define vector-set! mutable-vector-set!)
-(define vector-length mut-vec-len)
+; (define vector-length mut-vec-len)
 ; (define (make-vector n)
 ;   (apply mutable-vector (map (lambda (x) 0) (range 0 n))))
